Alpha version
Limited to a single user whose credentials cannot be changed. Mainly for testing releases.

+++ api/reddit/client.go
@@ -2,36 +2,48 @@ package reddit
 
 import (
 	"context"
-	"os"
+	"fmt"
 	"time"
 
 	"github.com/go-loremipsum/loremipsum"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
 	"github.com/vartanbeno/go-reddit/v2/reddit"
+	"github.com/wailsapp/wails/v2/pkg/runtime"
 )
 
+type Log struct {
+	Time    time.Time `json:"timestamp"`
+	Message string    `json:"message"`
+	Done    bool      `json:"done"`
+	Error   error     `json:"error"`
+}
+
 type Client struct {
+	ctx        context.Context
 	client     *reddit.Client
 	loremipsum *loremipsum.LoremIpsum
-	logger     zerolog.Logger
 	debug      bool
+	dryRun     bool
 }
 
 type NukeRequest struct {
-	Scheduled             bool
-	CronExpression        string
-	Posts                 bool
-	Comments              bool
-	MaxAge                int
-	UseMaxAge             bool
-	MaxScore              int
-	UseMaxScore           bool
-	ReplacementTextLength int
+	Scheduled             bool   `json:"scheduled"`
+	CronExpression        string `json:"cronExpression"`
+	Posts                 bool   `json:"posts"`
+	Comments              bool   `json:"comments"`
+	MaxAge                int    `json:"maxAge"`
+	UseMaxAge             bool   `json:"useMaxAge"`
+	MinScore              int    `json:"minScore"`
+	UseMinScore           bool   `json:"useMinScore"`
+	ReplacementTextLength int    `json:"replacementTextLength"`
+}
+
+type NukeResult struct {
+	CommentsDeleted int `json:"commentsDeleted"`
+	PostsDeleted    int `json:"postsDeleted"`
 }
 
 // NewClient creates a new client with the given credentials
-func NewClient(clientID, clientSecret, username, password string) (*Client, error) {
+func NewClient(ctx context.Context, clientID, clientSecret, username, password string, dryRun bool) (*Client, error) {
 	credentials := reddit.Credentials{ID: clientID, Secret: clientSecret, Username: username, Password: password}
 	client, err := reddit.NewClient(credentials)
 	if err != nil {
@@ -39,66 +51,79 @@ func NewClient(clientID, clientSecret, username, password string) (*Client, erro
 	}
 
 	return &Client{
+		ctx:        ctx,
 		client:     client,
 		loremipsum: loremipsum.New(),
-		logger:     log.Output(zerolog.ConsoleWriter{Out: os.Stderr}),
+		dryRun:     dryRun,
+		debug:      dryRun,
 	}, nil
 }
 
 // SetDebug sets the debug mode for the client
 func (c *Client) SetDebug(debug bool) {
 	c.debug = debug
-
-	if debug {
-		c.logger = c.logger.Level(zerolog.DebugLevel)
-	} else {
-		c.logger = c.logger.Level(zerolog.InfoLevel)
-	}
 }
 
-func (c *Client) Nuke(req NukeRequest) error {
+func (c *Client) Nuke(req NukeRequest, logChan chan<- Log) (NukeResult, error) {
+	result := NukeResult{}
+
 	if req.Posts {
-		err := c.EditAndDeleteAllUserPosts(req.ReplacementTextLength)
+		posts, err := c.EditAndDeleteAllUserPosts(req, logChan)
+		result.PostsDeleted = posts
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to delete posts")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to delete posts: %s", err.Error())
+			logChan <- Log{Time: time.Now(), Message: "Failed to delete posts", Done: true, Error: err}
+			return result, err
 		}
 	}
 
 	if req.Comments {
-		err := c.EditAndDeleteAllUserComments(req.ReplacementTextLength)
+		comments, err := c.EditAndDeleteAllUserComments(req, logChan)
+		result.CommentsDeleted = comments
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to delete comments")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to delete comments: %s", err.Error())
+			logChan <- Log{Time: time.Now(), Message: "Failed to delete comments", Done: true, Error: err}
+			return result, err
 		}
 	}
 
-	return nil
+	logChan <- Log{Time: time.Now(), Message: "Nuke complete", Done: true, Error: nil}
+
+	return result, nil
 }
 
 // EditAndDeleteAllUserPosts edits and deletes all posts for the authenticated user. It takes a parameter of count to set the number of words for the post.
-func (c *Client) EditAndDeleteAllUserPosts(count int) error {
-	posts, err := c.RetrieveAllUserPosts()
+func (c *Client) EditAndDeleteAllUserPosts(req NukeRequest, logChan chan<- Log) (int, error) {
+	posts, err := c.RetrieveAllUserPosts(logChan)
 	if err != nil {
-		c.logger.Error().Err(err).Msg("Failed to retrieve posts")
-		return err
+		runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to retrieve posts: %s", err.Error())
+		return 0, err
 	}
 
 	deleted := 0
 
 	for _, post := range posts {
-		err := c.EditPost(post, c.GenerateWords(count))
+		// Evaluate request and skip ones that match the criteria
+		if req.UseMaxAge && time.Since(post.Created.Time) < time.Duration(req.MaxAge*24)*time.Hour {
+			continue
+		}
+
+		if req.UseMinScore && post.Score > req.MinScore {
+			continue
+		}
+
+		err := c.EditPost(post, c.GenerateWords(req.ReplacementTextLength), logChan)
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to edit post")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to edit post: %s", err.Error())
+			return deleted, err
 		}
 
 		time.Sleep(1 * time.Second)
 
-		err = c.DeletePost(post)
+		err = c.DeletePost(post, logChan)
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to delete post")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to delete post: %s", err.Error())
+			return deleted, err
 		}
 
 		time.Sleep(1 * time.Second)
@@ -106,13 +131,13 @@ func (c *Client) EditAndDeleteAllUserPosts(count int) error {
 		deleted++
 	}
 
-	c.logger.Info().Int("PostsDeleted", deleted).Msg("All posts edited and deleted")
+	runtime.LogInfof(c.ctx, "[Reddit Client] All posts edited and deleted: %d", deleted)
 
-	return nil
+	return deleted, nil
 }
 
 // RetrieveAllUserPosts retrieves all posts for the authenticated user
-func (c *Client) RetrieveAllUserPosts() ([]*reddit.Post, error) {
+func (c *Client) RetrieveAllUserPosts(logChan chan<- Log) ([]*reddit.Post, error) {
 	var (
 		posts []*reddit.Post
 		after string
@@ -138,14 +163,16 @@ func (c *Client) RetrieveAllUserPosts() ([]*reddit.Post, error) {
 		secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
 		if c.debug {
-			for i, post := range pagePosts {
-				c.logger.Debug().Str("PostID", post.FullID).Int("index", i).Str("Created", post.Created.String()).Msg("Post fetched")
+			for _, post := range pagePosts {
+				runtime.LogDebugf(c.ctx, "[Reddit Client] Post fetched: %s", post.FullID)
+				logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Post fetched: %s", post.FullID), Done: false, Error: nil}
 			}
 		}
 
 		posts = append(posts, pagePosts...)
 
-		c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Int("PostsCount", len(posts)).Int("PagePosts", len(pagePosts)).Msg("Posts fetched")
+		runtime.LogDebugf(c.ctx, "[Reddit Client] Posts fetched: %d", len(posts))
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Fetched %d posts; total: %d", len(pagePosts), len(posts)), Done: false, Error: nil}
 
 		// less than the limit means we're done
 		if len(pagePosts) < 100 {
@@ -163,8 +190,14 @@ func (c *Client) RetrieveAllUserPosts() ([]*reddit.Post, error) {
 }
 
 // DeletePost deletes a single post by extract its ID
-func (c *Client) DeletePost(post *reddit.Post) error {
+func (c *Client) DeletePost(post *reddit.Post, logChan chan<- Log) error {
 	ctx := context.Background()
+
+	if c.dryRun {
+		runtime.LogInfof(c.ctx, "[Reddit Client] Dry run: Post would be deleted: %s", post.FullID)
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Dry run: Post would be deleted: %s", post.FullID), Done: false, Error: nil}
+		return nil
+	}
 
 	resp, err := c.client.Post.Delete(ctx, post.FullID)
 	if err != nil {
@@ -174,7 +207,7 @@ func (c *Client) DeletePost(post *reddit.Post) error {
 	remaining := resp.Header.Get("X-Ratelimit-Remaining")
 	secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
-	c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Msg("Post deleted")
+	runtime.LogDebugf(c.ctx, "[Reddit Client] Post deleted: %s", post.FullID)
 
 	err = evaluateLimit(remaining, secUntilReset)
 
@@ -182,8 +215,14 @@ func (c *Client) DeletePost(post *reddit.Post) error {
 }
 
 // EditPost edits a single post, setting the content to the given words
-func (c *Client) EditPost(post *reddit.Post, words string) error {
+func (c *Client) EditPost(post *reddit.Post, words string, logChan chan<- Log) error {
 	ctx := context.Background()
+
+	if c.dryRun {
+		runtime.LogInfof(c.ctx, "[Reddit Client] Dry run: Post would be edited: %s", post.FullID)
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Dry run: Post would be edited: %s", post.FullID), Done: false, Error: nil}
+		return nil
+	}
 
 	_, resp, err := c.client.Post.Edit(ctx, post.FullID, words)
 	if err != nil {
@@ -193,7 +232,7 @@ func (c *Client) EditPost(post *reddit.Post, words string) error {
 	remaining := resp.Header.Get("X-Ratelimit-Remaining")
 	secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
-	c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Msg("Post edited")
+	runtime.LogDebugf(c.ctx, "[Reddit Client] Post edited: %s", post.FullID)
 
 	// if we're out of requests, sleep
 	err = evaluateLimit(remaining, secUntilReset)
@@ -202,28 +241,37 @@ func (c *Client) EditPost(post *reddit.Post, words string) error {
 }
 
 // EditAndDeleteAllUserComments edits and deletes all comments for the authenticated user. It takes a parameter of count to set the number of words for the comment.
-func (c *Client) EditAndDeleteAllUserComments(count int) error {
-	comments, err := c.RetrieveAllUserComments()
+func (c *Client) EditAndDeleteAllUserComments(req NukeRequest, logChan chan<- Log) (int, error) {
+	comments, err := c.RetrieveAllUserComments(logChan)
 	if err != nil {
-		c.logger.Error().Err(err).Msg("Failed to retrieve comments")
-		return err
+		runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to retrieve comments: %s", err.Error())
+		return 0, err
 	}
 
 	deleted := 0
 
 	for _, comment := range comments {
-		err := c.EditComment(comment, c.GenerateWords(count))
+		// Evaluate request and skip ones that match the criteria
+		if req.UseMaxAge && time.Since(comment.Created.Time) < time.Duration(req.MaxAge*24)*time.Hour {
+			continue
+		}
+
+		if req.UseMinScore && comment.Score > req.MinScore {
+			continue
+		}
+
+		err := c.EditComment(comment, c.GenerateWords(req.ReplacementTextLength), logChan)
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to edit comment")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to edit comment: %s", err.Error())
+			return deleted, err
 		}
 
 		time.Sleep(1 * time.Second)
 
-		err = c.DeleteComment(comment)
+		err = c.DeleteComment(comment, logChan)
 		if err != nil {
-			c.logger.Error().Err(err).Msg("Failed to delete comment")
-			return err
+			runtime.LogErrorf(c.ctx, "[Reddit Client] Failed to delete comment: %s", err.Error())
+			return deleted, err
 		}
 
 		time.Sleep(1 * time.Second)
@@ -231,9 +279,10 @@ func (c *Client) EditAndDeleteAllUserComments(count int) error {
 		deleted++
 	}
 
-	c.logger.Info().Int("CommentsDeleted", deleted).Msg("All comments edited and deleted")
+	runtime.LogInfof(c.ctx, "[Reddit Client] All comments edited and deleted: %d", deleted)
+	logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("All comments edited and deleted: %d", deleted), Done: false, Error: nil}
 
-	return nil
+	return deleted, nil
 }
 
 // GenerateWords generates words using a LoremIpsum generator
@@ -242,8 +291,14 @@ func (c *Client) GenerateWords(quantity int) string {
 }
 
 // DeleteComment deletes a single comment by extract its ID
-func (c *Client) DeleteComment(comment *reddit.Comment) error {
+func (c *Client) DeleteComment(comment *reddit.Comment, logChan chan<- Log) error {
 	ctx := context.Background()
+
+	if c.dryRun {
+		runtime.LogInfof(c.ctx, "[Reddit Client] Dry run: Comment would be deleted: %s", comment.FullID)
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Dry run: Comment would be deleted: %s", comment.FullID), Done: false, Error: nil}
+		return nil
+	}
 
 	resp, err := c.client.Comment.Delete(ctx, comment.FullID)
 	if err != nil {
@@ -253,7 +308,8 @@ func (c *Client) DeleteComment(comment *reddit.Comment) error {
 	remaining := resp.Header.Get("X-Ratelimit-Remaining")
 	secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
-	c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Msg("Comment deleted")
+	runtime.LogDebugf(c.ctx, "[Reddit Client] Comment deleted: %s", comment.FullID)
+	logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Comment deleted: %s", comment.FullID), Done: false, Error: nil}
 
 	err = evaluateLimit(remaining, secUntilReset)
 
@@ -261,8 +317,14 @@ func (c *Client) DeleteComment(comment *reddit.Comment) error {
 }
 
 // EditComment edits a single comment, setting the content to the given words
-func (c *Client) EditComment(comment *reddit.Comment, words string) error {
+func (c *Client) EditComment(comment *reddit.Comment, words string, logChan chan<- Log) error {
 	ctx := context.Background()
+
+	if c.dryRun {
+		runtime.LogInfof(c.ctx, "[Reddit Client] Dry run: Comment would be edited: %s", comment.FullID)
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Dry run: Comment would be edited: %s", comment.FullID), Done: false, Error: nil}
+		return nil
+	}
 
 	_, resp, err := c.client.Comment.Edit(ctx, comment.FullID, words)
 	if err != nil {
@@ -272,7 +334,8 @@ func (c *Client) EditComment(comment *reddit.Comment, words string) error {
 	remaining := resp.Header.Get("X-Ratelimit-Remaining")
 	secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
-	c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Msg("Comment edited")
+	runtime.LogDebugf(c.ctx, "[Reddit Client] Comment edited: %s", comment.FullID)
+	logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Comment edited: %s", comment.FullID), Done: false, Error: nil}
 
 	// if we're out of requests, sleep
 	err = evaluateLimit(remaining, secUntilReset)
@@ -281,7 +344,7 @@ func (c *Client) EditComment(comment *reddit.Comment, words string) error {
 }
 
 // RetrieveAllUserComments retrieves all comments for the authenticated user
-func (c *Client) RetrieveAllUserComments() ([]*reddit.Comment, error) {
+func (c *Client) RetrieveAllUserComments(logChan chan<- Log) ([]*reddit.Comment, error) {
 	var (
 		comments []*reddit.Comment
 		after    string
@@ -307,14 +370,16 @@ func (c *Client) RetrieveAllUserComments() ([]*reddit.Comment, error) {
 		secUntilReset := resp.Header.Get("X-Ratelimit-Reset")
 
 		if c.debug {
-			for i, comment := range pageComments {
-				c.logger.Debug().Str("CommentID", comment.FullID).Int("index", i).Str("Created", comment.Created.String()).Msg("Comment fetched")
+			for _, comment := range pageComments {
+				runtime.LogDebugf(c.ctx, "[Reddit Client] Comment fetched: %s", comment.FullID)
+				logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Comment fetched: %s", comment.FullID), Done: false, Error: nil}
 			}
 		}
 
 		comments = append(comments, pageComments...)
 
-		c.logger.Debug().Str("RateLimitUsed", resp.Header.Get("X-Ratelimit-Used")).Str("RateLimitRemaining", remaining).Str("RateLimitReset", secUntilReset).Int("CommentsCount", len(comments)).Int("PageComments", len(pageComments)).Msg("Comments fetched")
+		runtime.LogDebugf(c.ctx, "[Reddit Client] Comments fetched: %d", len(comments))
+		logChan <- Log{Time: time.Now(), Message: fmt.Sprintf("Fetched %d comments; total: %d", len(pageComments), len(comments)), Done: false, Error: nil}
 
 		// less than the limit means we're done
 		if len(pageComments) < 100 {
